refactor(chainStep): move chainRec step functions and equals into shared.js
